Unit conversion and bug fixes
Added outlier reassignment to clustering to catch spread origins.

Changed pixsize to apply to coords vs eps, so coords and calculated eps will all be in microns

Fixed an issue with dunkin where mean(first) was producing a single value. If the first timepoint has only one point then that point is used as the centroid.